LoginPage js to tsx
